requirements.txt
google-generativeai
ipython
streamlit
pandas
matplotlib
requests
geopy
tqdm
folium
streamlit-folium
textblob
regex